// src/naruto/enums/NarutoOptions.enum.ts
export enum NarutoOptions {
  NAME = 'name',
  DESCRIPTION = 'description',
  PHOTO = 'photo',
  TITLES = 'titles',
  BIRTHDAY = 'birthday',
  STATUS = 'status',
  AGE = 'age',
  HEIGHT = 'height',
  WEIGHT = 'weight',
  CLASSIFICATION = 'classification',
  TEAM = 'team',
  PARTNER = 'partner',
  OCCUPATION = 'occupation',
  NINJA_RANK = 'ninja_rank',
  NINJA_REGISTRATION = 'ninja_registration',
  FAMILY = 'family',
  JUTSU = 'jutsu',
  NATURE_TYPE = 'nature_type',
  TOOLS = 'tools',
  QUOTES = 'quotes',
  SEX = 'sex',
  DEBUT = 'debut',
  VOICES = 'voices',
  KEKKEI_GENKAI = 'kekkei_genkai',
  AFFILIATION = 'affiliation',
  CLAN = 'clan'
}